requirements.txt
praw==7.7.1
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
nltk==3.8.1
textblob==0.17.1
scipy==1.11.4
tqdm==4.66.1
python-dotenv==1.0.0
plotly==5.18.0